chore: harden BLE payload handling, add CI checks, and update release docs
- Refactor BLE TX task path and fix transmit task filename typo
- Add centralized payload validation helper and enforce length checks in GATT handlers
- Remove stale GATT declaration and clean up sensor/task include layering
- Normalize line endings and apply source consistency cleanups
- Fix cleanup-introduced build issues (Sensor include, signed/unsigned compare in IMP path)
- Add sanity/smoke/warning-gate scripts under scripts/
- Add GitHub Actions workflow for firmware sanity + warning checks
- Expand docs (BLE protocol, build/flash rollback, release checklist, docs index, test plan)
- Update revision history with unreleased changes since last push

+++ firmware/src/sensors/EChem_Imp.cpp
@@ -3,7 +3,6 @@
 #include "HWConfig/constants.h"
 #include "drivers/ad5940_hal.h"
 #include "power/power.h"
-#include "sensors/Sensor.h"
 #include "util/debug_log.h"
 
 
@@ -528,7 +527,7 @@ void sensor::EChem_Imp::ISR(void) {
 AD5940Err sensor::EChem_Imp::updateRegisters(void) {
   if (config.NumOfData > 0) {
     config.FifoDataCount += getNumBytesAvailable() / 4;
-    if (config.FifoDataCount >= config.NumOfData) {
+    if (config.FifoDataCount >= static_cast<uint32_t>(config.NumOfData)) {
       AD5940_WUPTCtrl(bFALSE);
       return AD5940ERR_OK;
     }
